fix destination flow

See merge request development/client-projects/asia-pacific/indonesia/tti-tsel/telsel-tti2-android!64

+++ tti2sdk/src/main/java/com/timwe/tti2sdk/domain/DestinationsUseCase.kt
@@ -6,7 +6,7 @@ import com.timwe.tti2sdk.data.net.api.Results
 
 interface DestinationsUseCase {
 
-    suspend fun getCityInfo(cityId: Int): Results<Destination>
+    suspend fun getCityInfo(cityId: Long): Results<Destination>
 
     suspend fun saveCities(listCities: ListCities)
